// src/app/(auth-scope)/sign-up/components/AuthSection/schema.ts
import z, { email } from "zod";

export const AuthSchema = z.object({
    email: z.email("Insira um e-mail válido"),
    password: z.string().min(6, "A senha deve ter no mínimo 6 caracteres"),
    name: z.string().min(2, "O nome deve ter no mínimo 2 caracteres"),
})

export type AuthSchemaType = z.infer<typeof AuthSchema>;